+++ flutter/lib/main.dart
@@ -6,35 +6,16 @@ import 'package:firebase_analytics/observer.dart';
 import 'package:firebase_database/firebase_database.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 import 'flutter/localization.dart';
 import 'models/base/transaction.dart';
 import 'remote/error_reporting.dart';
 import 'views/decks_list/decks_list.dart';
-import 'views/helpers/sign_in_widget.dart';
 import 'views/onboarding/intro_view.dart';
 
-class App extends StatefulWidget {
+class App extends StatelessWidget {
   static final _analyticsNavigatorObserver =
       FirebaseAnalyticsObserver(analytics: FirebaseAnalytics());
-
-  @override
-  AppState createState() => AppState();
-}
-
-class AppState extends State<App> {
-  static const String _introPrefKey = 'is-intro-shown';
-  final Future<SharedPreferences> _prefs = SharedPreferences.getInstance();
-  bool _isIntroShown;
-
-  @override
-  void initState() {
-    super.initState();
-    _prefs.then((pref) {
-      _isIntroShown = pref.getBool(_introPrefKey);
-    });
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -55,21 +36,9 @@ class AppState extends State<App> {
       ],
       navigatorObservers: [App._analyticsNavigatorObserver],
       title: title,
-      // SignInWidget must be above Navigator to provide CurrentUserWidget.of().
-      builder: (context, child) {
-        if (_isIntroShown != null && _isIntroShown != false) {
-          return SignInWidget(child: child);
-        } else {
-          return IntroViewWidget(callback: () {
-            setState(() {
-              _prefs.then((pref) {
-                pref.setBool(_introPrefKey, true);
-                _isIntroShown = true;
-              });
-            });
-          });
-        }
-      },
+      // OnboardingViewWidget must be above Navigator
+      // to provide CurrentUserWidget.of().
+      builder: (context, child) => OnboardingViewWidget(child: child),
       theme:
           ThemeData(primarySwatch: Colors.green, accentColor: Colors.redAccent),
       home: DecksListPage(title: title),
